% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_looic_com_to_no_policy_change.R
\name{add_looic_com_to_no_policy_change}
\alias{add_looic_com_to_no_policy_change}
\title{Add LOOIC Comparision for a Model Without the Effect of the Change in Policy}
\usage{
add_looic_com_to_no_policy_change(model, reloo = TRUE, cores = 1)
}
\arguments{
\item{model}{A model object as produced by \code{best_fit_model}}

\item{reloo}{Logical, defaults to \code{TRUE}. Should problematic
observations be refit.}

\item{cores}{Numeric, defaults to 1. The number of cores to use.}
}
\value{
A list containing the fitted model object and a tibble containing the
comparision LOOIC scores and standard error
}
\description{
Add LOOIC Comparision for a Model Without the Effect of the Change in Policy
}
\examples{

}
