% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Model.R
\name{Base_Model}
\alias{Base_Model}
\title{Generic BRM model with fitting defaults set}
\usage{
Base_Model(Predictors = NULL, ModelFamily, Data = NULL,
  SetIter = NULL, SetBurn = NULL, SetCores = NULL, Chains = 4,
  SetThin = 1, Silent = TRUE, SaveModel = TRUE, AdaptDelta = NULL,
  MaxTree = NULL)
}
\arguments{
\item{Predictors}{A character vector of variables to use as features}

\item{ModelFamily}{A character string containing the model family to fit.}

\item{Data}{A dataframe of the data to fit to.}

\item{SetIter}{Numeric, define the number of iterations}

\item{SetBurn}{Numeric, define the number of burn in iterations}

\item{SetCores}{Numeric, set the number of cores to use.}

\item{Chains}{Numeric, set the number of chains to use for MCMC}

\item{SetThin}{Nuemeric, set the number of iterations to thin over.}

\item{Silent}{Logical, defaults to \code{TRUE}. Should the results be returned without progress messages.}

\item{SaveModel}{Logical, defaults to \code{TRUE}. Should the model output be saved}

\item{AdaptDelta}{Numeric, adaptive delta to use}

\item{MaxTree}{Numeric, the maximum number of trees.}
}
\value{
A \code{brm} fitted model.
}
\description{
A generic model used to define default \code{brms} settings to be used across all
models investigated.
}
\examples{

}
\seealso{
ModelConstruct
}
