% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_fit_results_table.R
\name{best_fit_results_table}
\alias{best_fit_results_table}
\title{Function to generate results table}
\usage{
best_fit_results_table(df, Scheme, Pop, sep = ", ",
  model_to_select = 1)
}
\arguments{
\item{df}{A dataframe of model results}

\item{Scheme}{A character string specifying the vaccination scheme to make the
table for.}

\item{Pop}{A character string indicating which population to make the results table for.}

\item{sep}{A character vector indicating the seperator used between the upper and lower credible intervals.
The default is ', '.}

\item{model_to_select}{Numeric, defaults to 1. Manually select which model to extract. Useful if several models
are nearly comparable.}
}
\value{
A list containing the best fitting model name, the model caption, the tidied effect size
output and the summarised model for inspection
}
\description{
Function to generate results table
}
\examples{

}
\seealso{
model_comparision
}
