% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_stan_DLLs.R
\name{capture_stan_DLLs}
\alias{capture_stan_DLLs}
\title{Function to print the number of stan models loaded as DLLs and return their names}
\usage{
capture_stan_DLLs(verbose = TRUE, StanFileNames = "file",
  MaxDLLs = 100)
}
\arguments{
\item{verbose}{Logical, defaults to \code{TRUE}. If \code{TRUE} the number of loaded stan
models are printed.}

\item{StanFileNames}{A character string identifying the Stan file names, defaults to file.}

\item{MaxDLLs}{Numeric, DLL cap. The maximum defaults to 100.}
}
\value{
A character vector of the currently loaded DLLs.
}
\description{
Function to print the number of stan models loaded as DLLs and return their names
}
\examples{

}
