% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_rate.R
\name{case_rate}
\alias{case_rate}
\title{Function to estimate the case rate for a given variable}
\usage{
case_rate(df, rate_for, strat_by = NULL, age_split = NULL,
  CoB_split = TRUE, Year_strat = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{rate_for}{A character string indicating the variables to estiamte the rate for.}

\item{strat_by}{A character string indicating the variables to stratify by.}

\item{age_split}{A character string, defaults to \code{NULL}. Specifies the age group to
stratify for.}

\item{CoB_split}{Logical, defaults to \code{TRUE}. Should the case rate be grouped by the country of
birth.}

\item{Year_strat}{Logical, defaults to \code{TRUE}. Should the case rate be grouped by year.}
}
\value{
A dataframe of case rates with upper and lower bounds
}
\description{
Estimates the case rate for a given variable, stratified by given variables,
dependant on age, year, agegrp2, and uk birth status. Missing data for the rate variable
is dropped.
}
\examples{

}
\seealso{
extract_case_counts
}
