% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_user_DLL.R
\name{clear_user_DLL}
\alias{clear_user_DLL}
\title{Clear Stan DLLs}
\usage{
clear_user_DLL(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, defaults to \code{TRUE}. If \code{TRUE} the number of loaded stan
models are printed.}
}
\value{
This functions returns \code{NULL}
}
\description{
A wrapper function around \code{capture_stan_DLLS} and \code{UnloadUserDLLs}.
Identifies the loaded Stan DLLs and unloads them.
}
\examples{

}
\seealso{
capture_stan_DLLs UnloadUserDLLs
}
