% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_conf.R
\name{epi_conf}
\alias{epi_conf}
\title{Estimate Incidence Rates per 100,000}
\usage{
epi_conf(Cases, Population)
}
\arguments{
\item{Cases}{Numeric vector of cases}

\item{Population}{Numeric vector of populations}
}
\value{
Estimated incidence rates with confidence intervals.
}
\description{
Estimate Incidence Rates per 100,000
}
\examples{

}
