% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_inf_ef.R
\name{extract_inc_ef}
\alias{extract_inc_ef}
\title{Extract incidence estimates}
\usage{
extract_inc_ef(df, model)
}
\arguments{
\item{df}{A dataframe containing a Model and IRR variables.}

\item{model}{A Character string identifying the model to extract}
}
\value{
Returns a vector of formatted IRRs formatted for publication
}
\description{
Extract incidence estimates
}
\examples{

}
