% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_models.R
\name{fit_stan_models}
\alias{fit_stan_models}
\title{Fit and summarise models}
\usage{
fit_stan_models(ModelReadyDataSpec, PreCompile = TRUE,
  BaseLineModelsNames = NULL, FileName = "ModelResults", SubFiles = 4,
  Folder = "results", Path = ".", ReRunAnalysis = FALSE, verbose = TRUE,
  Cores = NULL, Iterations = 10000, BurnIn = 5000, Alpha = 0.99,
  Treedepth = 10, UpdateAlpha = 0.99999, HandleDLLs = FALSE,
  UpdateTreedepth = 20)
}
\arguments{
\item{ModelReadyDataSpec}{A dataframe of data ready for modelling}

\item{PreCompile}{Logical, defaults to \code{TRUE}. Should a subset of models be precompiled
(specified with \code{BaseLineModelNames}) for the first vaccination scheme and population?}

\item{BaseLineModelsNames}{A character string of baseline model names.}

\item{FileName}{A character string to use as the file name for the results.
Defaults to ModelResults.}

\item{SubFiles}{Numeric, defaults to 4. The number of subfiles to split the results into. Results are
saved as \code{ModelResults1} etc. This must be smaller or equal to the number of models to be fitted.}

\item{Folder}{Character string to use as the folder name for the results.
Defaults to "results"}

\item{Path}{Character string specifying the path to the results folder.
Defaults to ".".}

\item{ReRunAnalysis}{Logical, default to \code{FALSE}. Should the analysis be rerun.}

\item{verbose}{Logical defaults to \code{FALSE}. Should progress messages be printed.}

\item{Cores}{Numeric, defaults to the number of cores available minus 1.
How many cores to use for the analysis.}

\item{Iterations}{Numeric, the number of iterations to use for the MCMC sampler. Defaults to
10000.}

\item{BurnIn}{Numeric, the number of iterations to use for the burn in. Defaults to 5000.}

\item{Alpha}{Numeric, set the adaptive alpha value for the MCMC. Defaults to 0.99}

\item{Treedepth}{Numeric, set the maximum tree depth. Defaults to 10.}

\item{UpdateAlpha}{Numeric, set an updated alpha for models that have failed to fit.
Defaults to 0.99999.}

\item{HandleDLLs}{Logical, defaults to \code{FALSE}. Enables tracking and unloading of Stan DLL
objects. If not reachng R's DLL cap then this is not required.}

\item{UpdateTreedepth}{Numeric, set the updated maximum tree depth for models that have failed to fit.}
}
\value{
A dataframe of summarised results for the models specified to be fitted in the input dataframe.
}
\description{
Fit and summarise models
}
\examples{

}
\seealso{
clear_user_DLL Base_Model summarise_model_fit
}
