% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_est_sd.R
\name{format_est_sd}
\alias{format_est_sd}
\title{Format estimate and standard deviation}
\usage{
format_est_sd(Estimate, SE, Digits, Scientific = FALSE)
}
\arguments{
\item{Estimate}{Numeric, the effect estimate.}

\item{SE}{Numeric, the standard error}

\item{Digits}{Numeric, the number of digits to round to.}

\item{Scientific}{Logical, defaults to \code{FALSE}. Should the results be
displayed in scientific format.}
}
\value{
A format effect estimate with standard errors rounded to the specified digits.
}
\description{
Format estimate and standard deviation
}
\examples{

}
