% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_model_summaries.R
\name{format_model_summaries}
\alias{format_model_summaries}
\title{Summarise models}
\usage{
format_model_summaries(Models, ModelSpec, sep = ", ")
}
\arguments{
\item{Models}{A dataframe of fitted models with summary information}

\item{ModelSpec}{A dataframe containing descriptive information about each model}

\item{sep}{A character vector indicating the seperator used between the upper and lower credible intervals.
The default is ', '.}
}
\value{
A dataframe of formatted model summaries, including updated descriptive information.
}
\description{
Summarise fitted models and bind to descriptive information.
}
\examples{

}
