% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justify_char_numeric.R
\name{justify_char_numeric}
\alias{justify_char_numeric}
\title{Function to build justification framework for pander}
\usage{
justify_char_numeric(ColNum = NULL, Character = NULL, Numeric = NULL)
}
\arguments{
\item{ColNum}{Numeric, the total number of columns}

\item{Character}{Numeric vector, the numeric position of character columns.}

\item{Numeric}{Numeric vector, the numeric position of numeric columns.}
}
\value{
A character string specifying the justification for a pander table.
}
\description{
Function to build justification framework for pander
}
\examples{

}
