% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparision.R
\name{model_comparison}
\alias{model_comparison}
\title{Make Model Comparision, Filters by Scheme, Country of Birth, and Convergance}
\usage{
model_comparison(DF, Scheme, Pop, Digits = 3)
}
\arguments{
\item{DF}{A dataframe of summarised model results.}

\item{Scheme}{A character string specifying the scheme to select models for.}

\item{Pop}{A character string speccifying the population to select models for.}

\item{Digits}{Numeric, the number of digits to round to.}
}
\value{
A summarised table comparing model results
}
\description{
Make Model Comparision, Filters by Scheme, Country of Birth, and Convergance
}
\examples{

}
\seealso{
format_est_sd
}
