% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelConstruct.R
\name{ModelConstruct}
\alias{ModelConstruct}
\title{Specify the Model Formula using Population as an Offset and Cases as the Outcome}
\usage{
ModelConstruct(Preds)
}
\arguments{
\item{Preds}{A character vector containing variable names.}
}
\value{
A character string containing the model formula
}
\description{
Specify the Model Formula using Population as an Offset and Cases as the Outcome
}
\examples{

}
