% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_inf_rep.R
\name{OR_inf_rep}
\alias{OR_inf_rep}
\title{Inf Replacement for Large Odds Ratios}
\usage{
OR_inf_rep(OR, MaxSize = 1e+05)
}
\arguments{
\item{OR}{Numeric, or character odds ratio}

\item{MaxSize}{Numeric, maximum size for odds ratio. Defaults to
100000}
}
\value{
A bounded effect estimate or Inf.
}
\description{
Inf Replacement for Large Odds Ratios
}
\examples{

}
