% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_round_inf.R
\name{pretty_round_inf}
\alias{pretty_round_inf}
\title{Pretty format after checking a result is smaller than a limit}
\usage{
pretty_round_inf(x, digits = 2, ...)
}
\arguments{
\item{x}{Numeric or character effect size estimate}

\item{digits}{Numeric defaults to 2. The number of digits to round to.}

\item{...}{Additional arguements to pass to \code{OR_inf_rep}.}
}
\value{
A prettty effect estimate bounded by some maximum size.
}
\description{
Pretty format after checking a result is smaller than a limit
}
\examples{

}
