% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cohort_vac_cam.R
\name{sim_cohort_vac_cam}
\alias{sim_cohort_vac_cam}
\title{Identify those that are eligible for a vaccination program and track age specific incidence rates.}
\usage{
sim_cohort_vac_cam(df, eligib_vector, campaign, policy_change_yr = 2005)
}
\arguments{
\item{df}{A dataframe}

\item{eligib_vector}{Numeric year identifying eligibility.}

\item{campaign}{A character string identifying the name of the campaign.}

\item{policy_change_yr}{Numeric year identifying the year of the policy change.}
}
\value{
A simulate cohort for a given vaccination program correctly formatted as a
retrospective cohort.
}
\description{
This function identifies those that are eligible for vaccination programs and tracks their age
specific incidence over years since vaccination. It does this by,
\itemize{
\item Produces a data set with year of eligibility as a variable
\item Does this for both the universal and targeted schemes
\item Adds a variables for years since entry to scheme - both universal and targetted
\item Adds a flag for before or after policy change (i.e 2005 switch)
}
}
\examples{

}
