% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_warnings.R
\name{store_warnings}
\alias{store_warnings}
\title{Detect Warnings and Store as a List}
\usage{
store_warnings(ListFits)
}
\arguments{
\item{ListFits}{List of fitted model objects}
}
\value{
A list of model warnings
}
\description{
Detect Warnings and Store as a List
}
\examples{

}
