% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_model_fit.R
\name{summarise_model_fit}
\alias{summarise_model_fit}
\title{Summarise model fit results}
\usage{
summarise_model_fit(ModelsWithFits, Cores = 1)
}
\arguments{
\item{ModelsWithFits}{Dataframe containg fitted model objects (stored as a Fit variable).}

\item{cores}{Numeric, the number of cores to use for parralisation (used by loo). Default
is 1.}
}
\value{
A datafrmae containing summarised model fit statistics.
}
\description{
Summarise model fit results
}
\examples{

}
