% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnloadUserDLLs.R
\name{UnloadUserDLLs}
\alias{UnloadUserDLLs}
\title{Unload Named DLLs}
\usage{
UnloadUserDLLs(VectorDLLNames, verbose = TRUE)
}
\arguments{
\item{VectorDLLNames}{A character vetor of DLL names.}

\item{verbose}{Logical defautls to \code{TRUE}. Should progress messages be printed.}
}
\value{
This functions returns \code{NULL}
}
\description{
Unload Named DLLs
}
\examples{

}
